合并代码自wzg

+++ app/knowledgebase/model/CorporationShare.php
@@ -252,8 +252,8 @@ class CorporationShare extends Base{
      * @param  int $comment_id 评论id
      * @return [type]             [description]
      */
-    public function delOneCommentById($comment_id){
-        return $this->model->table($this->dbprefix.'corporation_share_comment')->where(['id'=>$comment_id])->delete();
+    public function delOneCommentById($comment_id,$uid){
+        return $this->model->table($this->dbprefix.'corporation_share_comment')->where(['id'=>$comment_id,'replyer_id'=>$uid])->delete();
     }
 
 }
